Bug fix - calls to resolved scope/keys should return promise
